refactor: Move `ExecuteCommand` to `test` Package
Creating an internal `test` package to hold test utilities. Refactored
the `ExecuteCommand` function to a separate type so we can simulate
encoded data being received from stdin.
