+++ main_suite_test.go
@@ -2,14 +2,12 @@ package main_test
 
 import (
 	"context"
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
 	"testing"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+
+	"github.com/adambkaplan/kidbase10/internal/test"
 )
 
 func TestKidBase10(t *testing.T) {
@@ -17,37 +15,9 @@ func TestKidBase10(t *testing.T) {
 	RunSpecs(t, "kidbase10 CLI")
 }
 
-func ExecuteCommand(ctx context.Context, cmd string, args ...string) (out string, err error) {
-	toRun := exec.Command(cmd, args...)
-	// Commands referencing local executables will have path separators
-	// Strip these out when creating temp files
-	baseCmd := filepath.Base(cmd)
-
-	combinedOut, err := os.CreateTemp("", fmt.Sprintf("%s-combined-out-*.log", baseCmd))
-	if err != nil {
-		return out, err
-	}
-	defer func() {
-		rmErr := os.Remove(combinedOut.Name())
-		if err == nil {
-			err = rmErr
-		}
-	}()
-	toRun.Stdout = combinedOut
-	toRun.Stderr = combinedOut
-	runErr := toRun.Run()
-	outBytes, readErr := os.ReadFile(combinedOut.Name())
-	out = string(outBytes)
-	if runErr != nil {
-		err = fmt.Errorf("command failed with exit code %d, output: %s",
-			toRun.ProcessState.ExitCode(), out)
-	}
-	if readErr != nil && err == nil {
-		err = readErr
-		return
-	}
-
-	return
+func ExecuteCommand(ctx context.Context, cmd string, args ...string) (string, error) {
+	commandRunner := test.NewCommandRunner(ctx, cmd, args...)
+	return commandRunner.Execute()
 }
 
 var _ = BeforeSuite(func(ctx context.Context) {
